backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
crewai==0.80.0
chromadb==0.5.23
pymupdf==1.24.14
beautifulsoup4==4.12.3
python-dotenv==1.0.1
python-multipart==0.0.9